Implement a test to ensure linking is done correctly.
"Note that this function does not take a handle, and so it can be
called without opening a device"
From the FTDI documentation, this appears to be one of the only calls
we can actually test against without an FTDI device attached to the
machine.

Because Windows is linked dynamically, this would require a different test.